Update read_location_data
Added functionality for writing to KML, even though it doesn't write attributes... Still gotta figure that one out.  For now it's pretty simple to just write out to shp or geojson and then use something like QGIS to convert it.

Also moved the exception handling for output type to the main() function.  It's a lot easier to just stop it and kill it there before it breaks down later.